Updates to the configuration
Adds ability to just set an MKL_LIB environment variable to point to the
location of the mkl libs on your system.  Also, you can set PREFIX for
installing the compiled executables, as well as conditionally enable or
disable a PROFILER variable to enable google-pprof of your binaries.

Also, we don't need -lithe and -parlib explicitly stated as dependencies
of tbb.  Doing so messes with us when we want to simply switch over to
the native linux versions by setting/unsetting LD_LIBRARY_PATH